Moved ResultOrFailure to Packet module
Also export it from `Network.NTP.Client` since it is part of logging
trace.

Also use `fold` instead of `foldMap id`.
